pyproject.toml
[tool.poetry]
name = "whatsapp"
version = "0.1.0"
description = ""
authors = ["Iago Elias <iagoelifa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyzmq = "^23.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.14.5"
mypy = "^0.961"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
